Review comment incorporation
